Adding and altering variables.
After running the initial experiment with the three topologies lattice
cycle and round robin and going through the analysis we understood
that a few new variables had to be keep in track.

Cooperating rating will be used to categorise the strategies and
because the variable cooperating rating was not saved as a parameter
in the previous example we are adding in now.

Furthermore, we want the normalised score instead of the score. And
it was being kept for the neighborhood score. Thus the normalised
score was added. Also the average neighborhood score now is
calculated based on the normalised score.

Because we change the order of the columns in the pandas data frame
the data_structure.py exists. We need to add the aforementioned
parameters there as well.