Script and pbs for complete graph.
In this experiment we will treat the roun robin like a complete
graph. The code is similar to the small world and binomial
but the graph created is a complete. We have tested in the spatial
tournament tests that the results are the same. We do this to be able
to compare the results with the rest topologies.